finishedTables
have populated the tables on my end with the sql script i created. Should be able to run the scrip either in parts or the whole thing from query. Then run the population sql script to populate the teams and players, user is dependent on user input and followed is also user input for future